// sdk/typescript-sdk-fern-backup/src/plato/_generated/api/types/CreateVmResponse.ts
// This file was auto-generated by Fern from our API Definition.

export interface CreateVmResponse {
    /** Status */
    status: string;
    /** Timestamp */
    timestamp?: string;
    /** Correlation ID for tracking */
    correlation_id: string;
    /** Public URL for the VM */
    url: string;
    /** Public job ID */
    job_public_id: string;
    /** Job group ID */
    job_group_id: string;
}